requirements.txt
fastapi
uvicorn
motor
python-multipart
python-jose[cryptography]
pydantic
python-dotenv
httpx
pytest-asyncio